+++ worddeposit/Controllers/UserInfoTVC.swift
@@ -139,7 +139,7 @@ class UserInfoTVC: UITableViewController {
                         UserService.shared.auth.handleFireAuthError(error, viewController: self)
                         return
                     }
-                    showLoginVC(view: self.view)
+                    PresentVC.loginVC(from: self.view)
                 }
             }
         }))
